queue Part

+++ coffeeshop/src/main/java/com/coffeeshop/Controller/ShopController.java
@@ -1,8 +1,11 @@
 package com.coffeeshop.Controller;
 
+import com.coffeeshop.EntityClasses.QueueEntity;
 import com.coffeeshop.EntityClasses.ShopEntity;
 import com.coffeeshop.EntityClasses.UserEntity;
 import com.coffeeshop.Model.BaseRestResponse;
+import com.coffeeshop.Model.Request.CreateQueue;
+import com.coffeeshop.Repository.QueueRepository;
 import com.coffeeshop.Repository.UserRepository;
 import com.coffeeshop.Services.ShopService;
 import com.fasterxml.jackson.core.JsonProcessingException;
@@ -22,10 +25,13 @@ public class ShopController {
     @Autowired
     UserRepository userRepository;
 
+    @Autowired
+    QueueRepository queueRepository;
+
     @PostMapping(value = "/create")
     public @ResponseBody
     ResponseEntity<String> createShop(@RequestBody ShopEntity shopEntity,
-                                      @RequestHeader("Authorization") String Authorization){
+                                      @RequestHeader("Authorization") String Authorization) {
 
         HttpHeaders responseHeaders = new HttpHeaders();
         responseHeaders.setContentType(MediaType.APPLICATION_JSON);
@@ -33,8 +39,8 @@ public class ShopController {
         BaseRestResponse baseResponse = new BaseRestResponse();
         String responseJson = "";
         try {
-            boolean Auth=validateToken(Authorization);
-            if(!Auth){
+            boolean Auth = validateToken(Authorization);
+            if (!Auth) {
                 baseResponse.setError(true);
                 baseResponse.setData(null);
                 baseResponse.setMessage("Authentication Fail");
@@ -49,7 +55,7 @@ public class ShopController {
             responseJson = mapper.writeValueAsString(baseResponse);
             return new ResponseEntity<String>(responseJson, responseHeaders, HttpStatus.OK);
 
-        }catch (Exception ex){
+        } catch (Exception ex) {
             baseResponse.setError(true);
             baseResponse.setData(null);
             baseResponse.setMessage(ex.getMessage().toString());
@@ -81,13 +87,23 @@ public class ShopController {
 
     @PutMapping(value = "/update")
     public @ResponseBody
-    ResponseEntity<String> updateShop(@RequestBody ShopEntity shopEntity){
+    ResponseEntity<String> updateShop(@RequestBody ShopEntity shopEntity,
+                                      @RequestHeader("Authorization") String Authorization) {
         HttpHeaders responseHeaders = new HttpHeaders();
         responseHeaders.setContentType(MediaType.APPLICATION_JSON);
         ObjectMapper mapper = new ObjectMapper();
         BaseRestResponse baseResponse = new BaseRestResponse();
         String responseJson = "";
         try {
+            boolean Auth = validateToken(Authorization);
+            if (!Auth) {
+                baseResponse.setError(true);
+                baseResponse.setData(null);
+                baseResponse.setMessage("Authentication Fail");
+                responseJson = mapper.writeValueAsString(baseResponse);
+                return new ResponseEntity<String>(responseJson, responseHeaders, HttpStatus.UNAUTHORIZED);
+            }
+
             final ShopEntity entity = shopService.Update(shopEntity);
             baseResponse.setError(false);
             baseResponse.setData(entity);
@@ -96,7 +112,7 @@ public class ShopController {
             responseJson = mapper.writeValueAsString(baseResponse);
             return new ResponseEntity<String>(responseJson, responseHeaders, HttpStatus.OK);
 
-        }catch (Exception ex){
+        } catch (Exception ex) {
             baseResponse.setError(true);
             baseResponse.setData(null);
             baseResponse.setMessage(ex.getMessage().toString());
@@ -109,15 +125,26 @@ public class ShopController {
             return new ResponseEntity<String>(responseJson, responseHeaders, HttpStatus.INTERNAL_SERVER_ERROR);
         }
     }
+
     @DeleteMapping(value = "/delete")
     public @ResponseBody
-    ResponseEntity<String> deleteShop(@RequestBody ShopEntity shopEntity){
+    ResponseEntity<String> deleteShop(@RequestBody ShopEntity shopEntity,
+                                      @RequestHeader("Authorization") String Authorization) {
         HttpHeaders responseHeaders = new HttpHeaders();
         responseHeaders.setContentType(MediaType.APPLICATION_JSON);
         ObjectMapper mapper = new ObjectMapper();
         BaseRestResponse baseResponse = new BaseRestResponse();
         String responseJson = "";
         try {
+            boolean Auth = validateToken(Authorization);
+            if (!Auth) {
+                baseResponse.setError(true);
+                baseResponse.setData(null);
+                baseResponse.setMessage("Authentication Fail");
+                responseJson = mapper.writeValueAsString(baseResponse);
+                return new ResponseEntity<String>(responseJson, responseHeaders, HttpStatus.UNAUTHORIZED);
+            }
+
             final ShopEntity entity = shopService.Delete(shopEntity);
             baseResponse.setError(false);
             baseResponse.setData(entity);
@@ -126,7 +153,7 @@ public class ShopController {
             responseJson = mapper.writeValueAsString(baseResponse);
             return new ResponseEntity<String>(responseJson, responseHeaders, HttpStatus.OK);
 
-        }catch (Exception ex){
+        } catch (Exception ex) {
             baseResponse.setError(true);
             baseResponse.setData(null);
             baseResponse.setMessage(ex.getMessage().toString());
@@ -142,19 +169,29 @@ public class ShopController {
 
     @GetMapping(value = "/get/{id}")
     public @ResponseBody
-    ResponseEntity<String> deleteShop(@PathVariable("id") long id){
+    ResponseEntity<String> deleteShop(@PathVariable("id") long id,
+                                      @RequestHeader("Authorization") String Authorization) {
         HttpHeaders responseHeaders = new HttpHeaders();
         responseHeaders.setContentType(MediaType.APPLICATION_JSON);
         ObjectMapper mapper = new ObjectMapper();
         BaseRestResponse baseResponse = new BaseRestResponse();
         String responseJson = "";
         try {
+            boolean Auth = validateToken(Authorization);
+            if (!Auth) {
+                baseResponse.setError(true);
+                baseResponse.setData(null);
+                baseResponse.setMessage("Authentication Fail");
+                responseJson = mapper.writeValueAsString(baseResponse);
+                return new ResponseEntity<String>(responseJson, responseHeaders, HttpStatus.UNAUTHORIZED);
+            }
+
             final Optional<ShopEntity> entity = shopService.getShop(id);
             baseResponse.setError(false);
-            if(entity==null){
+            if (entity == null) {
                 baseResponse.setData(null);
                 baseResponse.setMessage("No Shops");
-            }else{
+            } else {
                 baseResponse.setData(entity.get());
                 baseResponse.setMessage("Shop Details");
             }
@@ -163,7 +200,7 @@ public class ShopController {
             responseJson = mapper.writeValueAsString(baseResponse);
             return new ResponseEntity<String>(responseJson, responseHeaders, HttpStatus.OK);
 
-        }catch (Exception ex){
+        } catch (Exception ex) {
             baseResponse.setError(false);
             baseResponse.setData(null);
             baseResponse.setMessage("No Shops");
@@ -178,7 +215,167 @@ public class ShopController {
     }
 
 
+    @PostMapping(value = "/queue/create")
+    public @ResponseBody
+    ResponseEntity<String> createQueue(@RequestBody CreateQueue CreateQueue,
+                                       @RequestHeader("Authorization") String Authorization) {
+        HttpHeaders responseHeaders = new HttpHeaders();
+        responseHeaders.setContentType(MediaType.APPLICATION_JSON);
+        ObjectMapper mapper = new ObjectMapper();
+        BaseRestResponse baseResponse = new BaseRestResponse();
+        String responseJson = "";
+        try {
+            boolean Auth = validateToken(Authorization);
+            if (!Auth) {
+                baseResponse.setError(true);
+                baseResponse.setData(null);
+                baseResponse.setMessage("Authentication Fail");
+                responseJson = mapper.writeValueAsString(baseResponse);
+                return new ResponseEntity<String>(responseJson, responseHeaders, HttpStatus.UNAUTHORIZED);
+            }
+            QueueEntity queueEntity = new QueueEntity();
+            Optional<ShopEntity> shop = shopService.getShop(CreateQueue.getShopid());
+            if (shop.get() != null) {
+                queueEntity.setShop(shop.get());
+                queueEntity.setMaxsize(CreateQueue.getMax());
+            } else {
+                baseResponse.setError(true);
+                baseResponse.setData(null);
+                baseResponse.setMessage("Shop is not valid");
+                baseResponse.setCode(500);
+                responseJson = mapper.writeValueAsString(baseResponse);
+                return new ResponseEntity<String>(responseJson, responseHeaders, HttpStatus.INTERNAL_SERVER_ERROR);
 
+            }
+            QueueEntity save = shopService.CreateQueue(queueEntity);
+            baseResponse.setError(false);
+            baseResponse.setData(save);
+            baseResponse.setMessage("Created Queue Successfully");
+            baseResponse.setCode(201);
+            responseJson = mapper.writeValueAsString(baseResponse);
+            return new ResponseEntity<String>(responseJson, responseHeaders, HttpStatus.OK);
+
+        } catch (Exception ex) {
+            baseResponse.setError(true);
+            baseResponse.setData(null);
+            baseResponse.setMessage(ex.getMessage().toString());
+            baseResponse.setCode(500);
+            try {
+                responseJson = mapper.writeValueAsString(baseResponse);
+            } catch (JsonProcessingException e) {
+                e.printStackTrace();
+            }
+            return new ResponseEntity<String>(responseJson, responseHeaders, HttpStatus.INTERNAL_SERVER_ERROR);
+        }
+    }
+
+    @PutMapping(value = "/queue/update")
+    public @ResponseBody
+    ResponseEntity<String> UpdateQueue(@RequestBody CreateQueue CreateQueue,
+                                       @RequestHeader("Authorization") String Authorization) {
+        HttpHeaders responseHeaders = new HttpHeaders();
+        responseHeaders.setContentType(MediaType.APPLICATION_JSON);
+        ObjectMapper mapper = new ObjectMapper();
+        BaseRestResponse baseResponse = new BaseRestResponse();
+        String responseJson = "";
+        try {
+            boolean Auth = validateToken(Authorization);
+            if (!Auth) {
+                baseResponse.setError(true);
+                baseResponse.setData(null);
+                baseResponse.setMessage("Authentication Fail");
+                responseJson = mapper.writeValueAsString(baseResponse);
+                return new ResponseEntity<String>(responseJson, responseHeaders, HttpStatus.UNAUTHORIZED);
+            }
+            QueueEntity queueEntity = new QueueEntity();
+            Optional<ShopEntity> shop = shopService.getShop(CreateQueue.getShopid());
+            if (shop.get() != null) {
+                queueEntity.setShop(shop.get());
+                queueEntity.setMaxsize(CreateQueue.getMax());
+            } else {
+                baseResponse.setError(true);
+                baseResponse.setData(null);
+                baseResponse.setMessage("Shop is not valid");
+                baseResponse.setCode(500);
+                responseJson = mapper.writeValueAsString(baseResponse);
+                return new ResponseEntity<String>(responseJson, responseHeaders, HttpStatus.INTERNAL_SERVER_ERROR);
+
+            }
+            QueueEntity save = shopService.UpdateQueue(queueEntity);
+            baseResponse.setError(false);
+            baseResponse.setData(save);
+            baseResponse.setMessage("Update Queue Successfully");
+            baseResponse.setCode(201);
+            responseJson = mapper.writeValueAsString(baseResponse);
+            return new ResponseEntity<String>(responseJson, responseHeaders, HttpStatus.OK);
+
+        } catch (Exception ex) {
+            baseResponse.setError(true);
+            baseResponse.setData(null);
+            baseResponse.setMessage(ex.getMessage().toString());
+            baseResponse.setCode(500);
+            try {
+                responseJson = mapper.writeValueAsString(baseResponse);
+            } catch (JsonProcessingException e) {
+                e.printStackTrace();
+            }
+            return new ResponseEntity<String>(responseJson, responseHeaders, HttpStatus.INTERNAL_SERVER_ERROR);
+        }
+    }
+
+    @DeleteMapping(value = "/queue/delete")
+    public @ResponseBody
+    ResponseEntity<String> DeleteQueue(@RequestBody CreateQueue CreateQueue,
+                                       @RequestHeader("Authorization") String Authorization) {
+        HttpHeaders responseHeaders = new HttpHeaders();
+        responseHeaders.setContentType(MediaType.APPLICATION_JSON);
+        ObjectMapper mapper = new ObjectMapper();
+        BaseRestResponse baseResponse = new BaseRestResponse();
+        String responseJson = "";
+        try {
+            boolean Auth = validateToken(Authorization);
+            if (!Auth) {
+                baseResponse.setError(true);
+                baseResponse.setData(null);
+                baseResponse.setMessage("Authentication Fail");
+                responseJson = mapper.writeValueAsString(baseResponse);
+                return new ResponseEntity<String>(responseJson, responseHeaders, HttpStatus.UNAUTHORIZED);
+            }
+            QueueEntity queueEntity = new QueueEntity();
+            Optional<ShopEntity> shop = shopService.getShop(CreateQueue.getShopid());
+            if (shop.get() != null) {
+                queueEntity.setShop(shop.get());
+                queueEntity.setMaxsize(CreateQueue.getMax());
+            } else {
+                baseResponse.setError(true);
+                baseResponse.setData(null);
+                baseResponse.setMessage("Shop is not valid");
+                baseResponse.setCode(500);
+                responseJson = mapper.writeValueAsString(baseResponse);
+                return new ResponseEntity<String>(responseJson, responseHeaders, HttpStatus.INTERNAL_SERVER_ERROR);
+
+            }
+            QueueEntity save = shopService.DeleteQueue(queueEntity);
+            baseResponse.setError(false);
+            baseResponse.setData(save);
+            baseResponse.setMessage("Delete Queue Successfully");
+            baseResponse.setCode(201);
+            responseJson = mapper.writeValueAsString(baseResponse);
+            return new ResponseEntity<String>(responseJson, responseHeaders, HttpStatus.OK);
+
+        } catch (Exception ex) {
+            baseResponse.setError(true);
+            baseResponse.setData(null);
+            baseResponse.setMessage(ex.getMessage().toString());
+            baseResponse.setCode(500);
+            try {
+                responseJson = mapper.writeValueAsString(baseResponse);
+            } catch (JsonProcessingException e) {
+                e.printStackTrace();
+            }
+            return new ResponseEntity<String>(responseJson, responseHeaders, HttpStatus.INTERNAL_SERVER_ERROR);
+        }
+    }
 
 
 }
